@@ -13,16 +13,20 @@
     "url": "http://github.com/nikgraf/belle.git"
   },
   "dependencies": {
-    "react": "=0.13.3",
+    "react": "=0.14.0",
+    "react-addons-linked-state-mixin": "^0.14.0",
+    "react-dom": "^0.14.0",
     "underscore": "=1.8.3"
   },
   "devDependencies": {
-    "babel-core": "^5.8.23",
+    "babel-core": "^5.8.25",
     "babel-loader": "^5.3.2",
     "express": "^4.13.3",
-    "node-libs-browser": "^0.5.2",
-    "webpack": "^1.12.1",
-    "webpack-dev-server": "^1.10.1"
+    "node-libs-browser": "^0.5.3",
+    "react-transform-catch-errors": "^1.0.0",
+    "redbox-react": "^1.1.1",
+    "webpack": "^1.12.2",
+    "webpack-dev-server": "^1.12.0"
   },
   "scripts": {
     "start": "node server.js"
