@@ -2,6 +2,7 @@ import React from 'react';
 import {Rating} from 'belle';
 import Code from './Code';
 import {propertyNameStyle, propertyDescriptionStyle} from './style';
+import LinkedStateMixin from 'react-addons-linked-state-mixin';
 
 const basicCodeExample = `<Rating defaultValue={3}></Rating>`;
 
@@ -9,17 +10,32 @@ const advanceCodeExample1 = `<Rating defaultValue={4} disabled></Rating>`;
 
 const advanceCodeExample2 = `<Rating defaultValue={4} character={'✪'}></Rating>`;
 
-const advanceCodeExample3 = `<Rating ref="rating" defaultValue={4}></Rating>
-<a onClick={ function() { this.refs.rating.resetValue(); }.bind(this); }
+const advanceCodeExample3 = `<Rating valueLink={ this.linkState('customRatingValue') }/>
+
+<a onClick={ this._resetRating }
    style={ {
      marginLeft: 20,
      position: 'relative',
      top: -5,
      textDecoration: 'underline',
      cursor: 'pointer'
-    } }>Reset</a>`;
+    } }>Reset</a>
+
+_resetRating() {
+  this.setState({
+    customRatingValue: undefined
+  });
+}`;
 
 export default React.createClass({
+
+  mixins: [LinkedStateMixin],
+
+  getInitialState() {
+    return {
+      customRatingValue: 3
+    };
+  },
 
   render() {
     return (<div>
@@ -32,7 +48,7 @@ export default React.createClass({
 
       <h3>Properties</h3>
 
-      <table>
+      <table><tbody>
         <tr>
           <td style={ propertyNameStyle }>
             valueLink
@@ -347,7 +363,7 @@ export default React.createClass({
             </p>
           </td>
         </tr>
-      </table>
+      </tbody></table>
 
       <p>
       Other supported properties include:
@@ -355,24 +371,6 @@ export default React.createClass({
         onMouseUp, onMouseEnter, onMouseMove, onMouseLeave, onTouchStart, onTouchMove, onTouchEnd, onTouchCancel, onFocus, onBlur, onClick,
         onKeyDown</span>
       </p>
-
-      <h3>Methods</h3>
-
-      <table>
-        <tr>
-          <td style={ propertyNameStyle }>
-            resetValue
-          </td>
-        </tr>
-        <tr>
-          <td style={ propertyDescriptionStyle }>
-            <p style={ {display: 'block'} }>
-              This method can be called to reset the Rating's value to undefined.<br/>
-              (Note: This method is not so much useful for controlled components. In those cases to set value to undefined props can be updated.)
-            </p>
-          </td>
-        </tr>
-      </table>
 
       <h3>More Examples</h3>
 
@@ -384,9 +382,10 @@ export default React.createClass({
       <Rating defaultValue={4} character={'✪'}/>
       <Code value= { advanceCodeExample2 } style={ {marginTop: 20, marginBottom: 40} } />
 
-      <h4>Rating with a Reset Link</h4>
-      <Rating ref="rating" defaultValue={4}/>
-      <a onClick={ (() => { this.refs.rating.resetValue(); }).bind(this) }
+      <h4>Controlled Rating Component with a Reset Link</h4>
+      <p>Reset rating functionality can be implemented using controlled rating component like this:</p>
+      <Rating valueLink={ this.linkState('customRatingValue') }/>
+      <a onClick={ this._resetRating }
          style={ {
            marginLeft: 20,
            position: 'relative',
@@ -396,5 +395,11 @@ export default React.createClass({
           } }>Reset</a>
       <Code value= { advanceCodeExample3 } style={ {marginTop: 20} } />
     </div>);
+  },
+
+  _resetRating() {
+    this.setState({
+      customRatingValue: undefined
+    });
   }
 });
