@@ -3,9 +3,8 @@
 jest.dontMock('../components/ComboBox');
 jest.dontMock('../components/Option');
 
-import {extend} from 'underscore';
-import React from 'react/addons';
-const TestUtils = React.addons.TestUtils;
+import React from 'react';
+import TestUtils from 'react-addons-test-utils';
 
 // Babel would move an import in front of the jest.dontMock. That's why require
 // is used instead of import.
@@ -125,7 +124,8 @@ describe('ComboBox', () => {
     );
 
     const selectedAreaNode = TestUtils.scryRenderedDOMComponentsWithTag(combobox, 'input')[1];
-    expect(selectedAreaNode.props.style.cursor).toBe('cross');
+    expect(selectedAreaNode.hasAttribute('style')).toBeTruthy();
+    expect(selectedAreaNode.getAttribute('style').indexOf('cursor:cross') > -1).toBeTruthy();
   });
 
   describe('updating props', () => {
@@ -141,7 +141,10 @@ describe('ComboBox', () => {
     });
 
     it('should update it\'s state in case value is provided', () => {
-      const properties = extend({}, combobox.props, {value: 'vienna'});
+      const properties = {
+        ...combobox.props,
+        value: 'vienna'
+      };
       combobox.componentWillReceiveProps(properties);
 
       expect(combobox.state.inputValue).toBe('vienna');
@@ -154,7 +157,10 @@ describe('ComboBox', () => {
         value: 'vienna'
       };
 
-      const properties = extend({}, combobox.props, {valueLink: valueLink});
+      const properties = {
+        ...combobox.props,
+        valueLink: valueLink
+      };
       combobox.componentWillReceiveProps(properties);
 
       expect(combobox.state.inputValue).toBe('vienna');
