@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import {extend, omit} from '../utils/helpers';
+import {omit} from '../utils/helpers';
 import style from '../style/placeholder';
 
 /**
@@ -48,9 +48,16 @@ export default class Placeholder extends Component {
   }
 
   render() {
-    let computedStyle = extend({}, style.style, this.props.style);
+    let computedStyle = {
+      ...style.style,
+      ...this.props.style
+    };
     if (this.props._isDisabled) {
-      computedStyle = extend(computedStyle, style.disabledStyle, this.props.disabledStyle);
+      computedStyle = {
+        ...computedStyle,
+        ...style.disabledStyle,
+        ...this.props.disabledStyle
+      };
     }
 
     return (
